refactors packages, class names and formatting
