--- sprinboot.restservice/src/main/java/poc/mcastro/sprinboot/restservice/model/Status.java
@@ -1,5 +1,0 @@
-package poc.mcastro.sprinboot.restservice.model;
-
-public enum Status {
-	APPROVED, REJECTED, FAILED;
-}
