--- sprinboot.restservice/src/main/java/poc/mcastro/sprinboot/restservice/model/TransactionType.java
@@ -1,5 +1,0 @@
-package poc.mcastro.sprinboot.restservice.model;
-
-public enum TransactionType {
-	CREDIT, DEBIT, REQUEST;
-}
