+++ sprinboot.restservice/src/main/java/poc/mcastro/sprinboot/restservice/money/core/DataNotFoundException.java
@@ -1,0 +1,14 @@
+package poc.mcastro.sprinboot.restservice.money.core;
+
+public class DataNotFoundException extends RuntimeException {
+    private String message;
+
+    public DataNotFoundException(String message) {
+        super(message);
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+}
