+++ sprinboot.restservice/src/main/java/poc/mcastro/sprinboot/restservice/money/money/account/GetMoneyTO.java
@@ -1,0 +1,35 @@
+package poc.mcastro.sprinboot.restservice.money.money.account;
+
+import java.io.Serializable;
+import java.math.BigDecimal;
+
+public class GetMoneyTO implements Serializable {
+    private AccountTO accountTO;
+    private BigDecimal ammount;
+    private String reason;
+
+    public AccountTO getAccountTO() {
+        return accountTO;
+    }
+
+    public void setAccountTO(AccountTO accountTO) {
+        this.accountTO = accountTO;
+    }
+
+    public BigDecimal getAmmount() {
+        return ammount;
+    }
+
+    public void setAmmount(BigDecimal ammount) {
+        this.ammount = ammount;
+    }
+
+    public String getReason() {
+        return reason;
+    }
+
+    public void setReason(String reason) {
+        this.reason = reason;
+    }
+
+}
