+++ sprinboot.restservice/src/main/java/poc/mcastro/sprinboot/restservice/money/money/transaction/TransactionRepository.java
@@ -1,13 +1,11 @@
-package poc.mcastro.sprinboot.restservice.repository;
-
-import java.util.List;
+package poc.mcastro.sprinboot.restservice.money.money.transaction;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import poc.mcastro.sprinboot.restservice.model.TransactionTO;
+import java.util.List;
 
 @Repository
 public interface TransactionRepository extends MongoRepository<TransactionTO, String> {
-	public List<TransactionTO> findByStatus(String status);
+    public List<TransactionTO> findByStatus(String status);
 }
