+++ sprinboot.restservice/src/main/java/poc/mcastro/sprinboot/restservice/money/money/transaction/TransactionType.java
@@ -1,0 +1,5 @@
+package poc.mcastro.sprinboot.restservice.money.money.transaction;
+
+public enum TransactionType {
+    CREDIT, DEBIT, REQUEST
+}
